only initialize store for current instance
