+++ examples/simple/index.js
@@ -1,30 +1,7 @@
 import Vue from 'vue'
-import zerotwo, { store, computed } from '../../src'
 import App from './App.vue'
-
-Vue.use(zerotwo)
-
-@store
-class MyStore {
-  state = {
-    count: 0
-  }
-
-  increment = () => this.state.count++
-
-  @computed
-  doubleCount() {
-    console.log(this)
-    return this.state.count * 2
-  }
-
-  random() {
-    this.doubleCount
-  }
-}
 
 new Vue({
   el: '#app',
-  store: MyStore,
   render: h => h(App)
 })
