--- src/StoreConsumer.js
@@ -1,9 +1,0 @@
-export default {
-  name: 'StoreConsumer',
-
-  functional: true,
-
-  render(h, ctx) {
-    return ctx.data.scopedSlots.default(ctx.parent.$store)
-  }
-}
